DENESHBN-16 Improved Audit utils

+++ modules/apps/portal-security-audit/portal-security-audit-event-generators-user-management/src/main/java/com/liferay/portal/security/audit/event/generators/user/management/internal/model/listener/RoleModelListener.java
@@ -102,7 +102,7 @@ public class RoleModelListener extends BaseModelListener<Role> {
 			).put(
 				"associationName", _auditMessageRoleAssociationUtil.getName(associationClassName)
 			).put(
-				"associationValue", _auditMessageRoleAssociationUtil.getValue(associationClassName, associationClassP)
+				"associationValue", _auditMessageRoleAssociationUtil.getValue(associationClassName, (long)associationClassP)
 			).put(
 				"associationClassName", associationClassName
 			);
